Datatype change

+++ DomainModel/Minutely.cs
@@ -8,6 +8,6 @@ namespace DomainModel
         public int Dt { get; set; } 
 
         [JsonProperty("precipitation")]
-        public int Precipitation { get; set; } 
+        public double Precipitation { get; set; } 
     }
 }
